+++ src/Trarizon.Library.Collections/HISTORY.md
@@ -4,6 +4,6 @@
 
 - Reimplement `PrefixTree(Dictionary)` with node, the original implementation is renamed to `ContiguousPrefixTree(Dicitionary)` and marked as experimental
 - Add `TraAlgorithm.LevenshteinDistance`
-- Add `IEnumerable.Intersperse`, `Interleave`, `RepeatInterleave`
+- Add `IEnumerable.Intersperse`, `Interleave`, `RepeatInterleave`, `ChunkBy`
 - Add `TraCollection.Singleton`
 - Remove `IEnumerable.LookAhead`